backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
email-validator==2.3.0
PyJWT==2.8.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23